// src/app/shared/store/models/claim.models.ts
export interface ClaimBannerField {
  name: string;
  label?: string;
  elementName?: string;
  value?: any;
  order?: number;
  showWhileCollapsed?: boolean;
  type?: string;
  width?: number;
  editable?: boolean;
}
// ...
export interface Apportionment {
  apportionmentFlag: boolean;
  apportionmentPercent: number;
  claimantPercent: number;
  otherApportionmentPercent: number;
  otherApportionmentDescription: string;
}